fix: rename build key and organize imports
